requirements.txt
# Core dependencies for Antigravity Agent Factory
# Install with: pip install -r requirements.txt
#
# For development dependencies (testing, linting), see requirements-dev.txt

# =============================================================================
# Core Dependencies
# =============================================================================

# YAML support for configuration files
pyyaml==6.0.3

# Template engine for code generation
Jinja2==3.1.5

# Async HTTP client for adapters
aiohttp==3.11.12

# Schema validation for JSON files
jsonschema==4.26.0

# =============================================================================
# Memory System (Inductive Learning)
# =============================================================================

# Local embeddings for semantic memory (no API key required)
# Downloads ~80MB model on first use
sentence-transformers==5.2.2

# Vector database client for memory store
# Note: mcp-server-qdrant removed as we use custom RAG MCP
qdrant-client==1.16.2

# Local embeddings optimization
fastembed==0.7.4


# NumPy (required by sentence-transformers and qdrant-client)
numpy==1.26.4; python_version < "3.13"
numpy>=2.1.0; python_version >= "3.13"

# =============================================================================
# Web / API
# =============================================================================

# FastAPI backend for IDX UI
fastapi>=0.115.0
uvicorn>=0.34.0

# =============================================================================
# AI & Agents
# =============================================================================

# LLM provider abstraction (Gemini via aisuite)
aisuite[google]>=0.1.14

# LangChain Ecosystem
langchain==1.2.10
langchain-classic==1.0.1
langchain-core==1.2.11
langchain-community==0.4.1
langchain-google-genai==4.2.1
langgraph==1.0.10
langchain-huggingface==1.2.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0

# PDF Processing
pymupdf==1.27.1

# Data Validation
pydantic==2.12.5
langchain-mcp-adapters>=0.2.1

# LangSmith tracing
langsmith>=0.3.0
